+++ app/Filament/Resources/StockMovementResource.php
@@ -35,10 +35,7 @@ class StockMovementResource extends BaseResource
     {
         return $schema
             ->components([
-                Forms\Components\Select::make('customer_id')
-                    ->relationship('customer', 'company_name')
-                    ->searchable()
-                    ->required(),
+                static::customerIdField(),
                 Forms\Components\TextInput::make('movement_no')->required()->maxLength(100),
                 Forms\Components\Select::make('product_id')
                     ->relationship('product', 'product_name')