+++ app/Models/Import.php
@@ -2,12 +2,13 @@
 
 namespace App\Models;
 
+use App\Models\Concerns\BelongsToCustomer;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Import extends Model
 {
-    use HasFactory;
+    use BelongsToCustomer, HasFactory;
 
     protected $fillable = [
         'customer_id',